Better Solution
Instead of grouping the combinations to avoid duplicates I created a decision tree which branches off into one or more instances of the initial candidate or the next candidate. This solution was vastly better in both space and time than the other. Complexity is O(2^target)

How would you do this if the numbers could be negative?